update ft_is_sorted(): changed the algorithm
Был изменен алгортм проверки стека на отсортированность. Теперь
главным критерием отсортированности последовательности является
возрастающая разница между соседними элементами, начиная с элемента
со значением 0 и не учитывая разницы между последним (length - 1) и
первым (0) элементами.Функция позволяет определить псевдосортиро-
ванные массивы, для сортировки которых требуется лишь прокрутка стека.

+++ algorithms/ft_merge_sort.c
@@ -2,8 +2,8 @@
 
 static void	ft_split_stack(t_stack **stack_a, t_stack **stack_b, t_status *status)
 {
-	int	i;
-	int	count;
+	int					i;
+	unsigned long long	count;
 
 	i = 0;
 	count = (status->max - status->min) / 2;
@@ -82,7 +82,6 @@ static void	ft_merge_stacks(t_stack **stack_a, t_stack **stack_b)
  * Это усложнило бы код в разы.
  */
 
-//дописать для 6 элементов! А контроль количества до 4 элементав будет снаружи, внутри будет до 6 включительно
 void	ft_merge_sort(t_stack **stack_a, t_stack **stack_b, t_status *status)
 {
 	ft_split_stack(stack_a, stack_b, status);
